NotifyIdentityService removed
