+++ Sources/WalletConnectNotify/Client/Wallet/NotifyClient.swift
@@ -22,8 +22,9 @@ public class NotifyClient {
 
     public let logger: ConsoleLogging
 
+    private let keyserverURL: URL
     private let pushClient: PushClient
-    private let identityService: NotifyIdentityService
+    private let identityClient: IdentityClient
     private let notifyStorage: NotifyStorage
     private let notifyAccountProvider: NotifyAccountProvider
     private let notifyMessageSubscriber: NotifyMessageSubscriber
@@ -38,8 +39,9 @@ public class NotifyClient {
     private let subscriptionWatcher: SubscriptionWatcher
 
     init(logger: ConsoleLogging,
+         keyserverURL: URL,
          kms: KeyManagementServiceProtocol,
-         identityService: NotifyIdentityService,
+         identityClient: IdentityClient,
          pushClient: PushClient,
          notifyMessageSubscriber: NotifyMessageSubscriber,
          notifyStorage: NotifyStorage,
@@ -57,8 +59,9 @@ public class NotifyClient {
          subscriptionWatcher: SubscriptionWatcher
     ) {
         self.logger = logger
+        self.keyserverURL = keyserverURL
         self.pushClient = pushClient
-        self.identityService = identityService
+        self.identityClient = identityClient
         self.notifyMessageSubscriber = notifyMessageSubscriber
         self.notifyStorage = notifyStorage
         self.deleteNotifySubscriptionRequester = deleteNotifySubscriptionRequester
@@ -76,17 +79,22 @@ public class NotifyClient {
     }
 
     public func prepareRegistration(account: Account, domain: String, allApps: Bool = true) async throws -> IdentityRegistrationParams {
-        return try await identityService.prepareRegistration(account: account, domain: domain, allApps: allApps)
+        return try await identityClient.prepareRegistration(
+            account: account,
+            domain: domain,
+            statement: makeStatement(allApps: allApps),
+            resources: [keyserverURL.absoluteString]
+        )
     }
 
     public func register(params: IdentityRegistrationParams, signature: CacaoSignature) async throws {
-        try await identityService.register(params: params, signature: signature)
+        try await identityClient.register(params: params, signature: signature)
         notifyAccountProvider.setAccount(try params.account)
         try await subscriptionWatcher.start()
     }
 
     public func unregister(account: Account) async throws {
-        try await identityService.unregister(account: account)
+        try await identityClient.unregister(account: account)
         notifyWatcherAgreementKeysProvider.removeAgreement(account: account)
         try notifyStorage.clearDatabase(account: account)
         notifyAccountProvider.logout()
@@ -126,7 +134,7 @@ public class NotifyClient {
     }
 
     public func isIdentityRegistered(account: Account) -> Bool {
-        return identityService.isIdentityRegistered(account: account)
+        return identityClient.isIdentityRegistered(account: account)
     }
 
     public func subscriptionsPublisher(account: Account) -> AnyPublisher<[NotifySubscription], Never> {
@@ -135,6 +143,18 @@ public class NotifyClient {
 
     public func messagesPublisher(topic: String) -> AnyPublisher<[NotifyMessageRecord], Never> {
         return notifyStorage.messagesPublisher(topic: topic)
+    }
+}
+
+private extension NotifyClient {
+
+    func makeStatement(allApps: Bool) -> String {
+        switch allApps {
+        case false:
+            return "I further authorize this app to send me notifications. Read more at https://walletconnect.com/notifications"
+        case true:
+            return "I further authorize this app to view and manage my notifications for ALL apps. Read more at https://walletconnect.com/notifications"
+        }
     }
 }
